# backend/schemata/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Test Observer",
    "description": "Test Observer API (see https://github.com/canonical/test_observer)",
    "license": {
      "name": "GNU Affero General Public License v3",
      "url": "https://raw.githubusercontent.com/canonical/test_observer/refs/heads/main/backend/LICENSE"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/v1/test-cases/reported-issues": {
      "get": {
        "tags": [
          "test-cases"
        ],
        "summary": "Get Reported Issues",
        "operationId": "get_reported_issues_v1_test_cases_reported_issues_get",
        "parameters": [
          {
            "name": "template_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Template Id"
            }
          },
          {
            "name": "case_name",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Case Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TestReportedIssueResponse"
                  },
                  "title": "Response Get Reported Issues V1 Test Cases Reported Issues Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "x-permissions": [
          "view_test_case_reported_issue"
        ]
      },
      "post": {
        "tags": [
          "test-cases"
        ],
        "summary": "Create Reported Issue",
        "operationId": "create_reported_issue_v1_test_cases_reported_issues_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TestReportedIssueRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TestReportedIssueResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "x-permissions": [
          "change_test_case_reported_issue"
        ]
      }
    },
    "/v1/test-cases/reported-issues/{issue_id}": {
      "put": {
        "tags": [
          "test-cases"
        ],
        "summary": "Update Reported Issue",
        "operationId": "update_reported_issue_v1_test_cases_reported_issues__issue_id__put",
        "parameters": [
          {
            "name": "issue_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Issue Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TestReportedIssueRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TestReportedIssueResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "x-permissions": [
          "change_test_case_reported_issue"
        ]
      },
      "delete": {
        "tags": [
          "test-cases"
        ],
        "summary": "Delete Reported Issue",
        "operationId": "delete_reported_issue_v1_test_cases_reported_issues__issue_id__delete",
        "parameters": [
          {
            "name": "issue_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Issue Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "x-permissions": [
          "change_test_case_reported_issue"
        ]
      }
    },
...
    "/v1/environments/reported-issues/{issue_id}": {
      "put": {
        "tags": [
          "environments"
        ],
        "summary": "Update Reported Issue",
        "operationId": "update_reported_issue_v1_environments_reported_issues__issue_id__put",
        "parameters": [
          {
            "name": "issue_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Issue Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EnvironmentReportedIssueRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnvironmentReportedIssueResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "x-permissions": [
          "change_environment_reported_issue"
        ]
      },
      "delete": {
        "tags": [
          "environments"
        ],
        "summary": "Delete Reported Issue",
        "operationId": "delete_reported_issue_v1_environments_reported_issues__issue_id__delete",
        "parameters": [
          {
            "name": "issue_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Issue Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "x-permissions": [
          "change_environment_reported_issue"
        ]
      }
    },
...
    "/v1/issues/{issue_id}": {
      "get": {
        "tags": [
          "issues"
        ],
        "summary": "Get Issue",
        "operationId": "get_issue_v1_issues__issue_id__get",
        "parameters": [
          {
            "name": "issue_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Issue Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IssueResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "x-permissions": [
          "view_issue"
        ]
      },
      "patch": {
        "tags": [
          "issues"
        ],
        "summary": "Patch Issue",
        "operationId": "patch_issue_v1_issues__issue_id__patch",
        "parameters": [
          {
            "name": "issue_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Issue Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IssuePatchRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IssueResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "x-permissions": [
          "change_issue"
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "EnvironmentReportedIssueRequest": {
        "properties": {
          "environment_name": {
            "type": "string",
            "title": "Environment Name"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "url": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 2083,
                "minLength": 1,
                "format": "uri"
              },
              {
                "type": "null"
              }
            ],
            "title": "Url"
          },
          "is_confirmed": {
            "type": "boolean",
            "title": "Is Confirmed"
          }
        },
        "type": "object",
        "required": [
          "environment_name",
          "description",
          "is_confirmed"
        ],
        "title": "EnvironmentReportedIssueRequest"
      },
      "EnvironmentReportedIssueResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "environment_name": {
            "type": "string",
            "title": "Environment Name"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "url": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 2083,
                "minLength": 1,
                "format": "uri"
              },
              {
                "type": "null"
              }
            ],
            "title": "Url"
          },
          "is_confirmed": {
            "type": "boolean",
            "title": "Is Confirmed"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "environment_name",
          "description",
          "url",
          "is_confirmed",
          "created_at",
          "updated_at"
        ],
        "title": "EnvironmentReportedIssueResponse"
      },
...
      "ExecutionMetadata": {
        "additionalProperties": {
          "items": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "description": "Execution Metadata Value"
          },
          "type": "array"
        },
        "propertyNames": {
          "maxLength": 200,
          "minLength": 1,
          "description": "Execution Metadata Category"
        },
        "type": "object",
        "title": "Execution Metadata",
        "description": "A mapping of string categories to lists of string values.",
        "examples": [
          {
            "charm": [
              "postgresql-k8s",
              "mysql-k8s"
            ],
            "charm:postgresql-k8s:revision": [
              "495"
            ]
          }
        ]
      },
...
      "FamilyName": {
        "type": "string",
        "enum": [
          "snap",
          "deb",
          "charm",
          "image"
        ],
        "title": "FamilyName"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "IssuePatchRequest": {
        "properties": {
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "status": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/IssueStatus"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "title": "IssuePatchRequest"
      },
...
      "IssueResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "source": {
            "$ref": "#/components/schemas/IssueSource"
          },
          "project": {
            "type": "string",
            "title": "Project"
          },
          "key": {
            "type": "string",
            "title": "Key"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "status": {
            "$ref": "#/components/schemas/IssueStatus"
          },
          "url": {
            "type": "string",
            "maxLength": 2083,
            "minLength": 1,
            "format": "uri",
            "title": "Url"
          },
          "labels": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Labels"
          },
          "attachment_rules": {
            "items": {
              "$ref": "#/components/schemas/MinimalIssueTestResultAttachmentRuleResponse"
            },
            "type": "array",
            "title": "Attachment Rules"
          }
        },
        "type": "object",
        "required": [
          "id",
          "source",
          "project",
          "key",
          "title",
          "status",
          "url",
          "attachment_rules"
        ],
        "title": "IssueResponse"
      },
      "IssueSource": {
        "type": "string",
        "enum": [
          "jira",
          "github",
          "launchpad"
        ],
        "title": "IssueSource"
      },
      "IssueStatus": {
        "type": "string",
        "enum": [
          "unknown",
          "open",
          "closed"
        ],
        "title": "IssueStatus"
      },
...
      "MinimalIssueTestResultAttachmentRuleResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "enabled": {
            "type": "boolean",
            "title": "Enabled"
          },
          "families": {
            "items": {
              "$ref": "#/components/schemas/FamilyName"
            },
            "type": "array",
            "title": "Families"
          },
          "environment_names": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Environment Names"
          },
          "test_case_names": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Test Case Names"
          },
          "template_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Template Ids"
          },
          "test_result_statuses": {
            "items": {
              "$ref": "#/components/schemas/TestResultStatus"
            },
            "type": "array",
            "title": "Test Result Statuses"
          },
          "execution_metadata": {
            "$ref": "#/components/schemas/ExecutionMetadata"
          }
        },
        "type": "object",
        "required": [
          "id",
          "enabled",
          "families",
          "environment_names",
          "test_case_names",
          "template_ids",
          "test_result_statuses",
          "execution_metadata"
        ],
        "title": "MinimalIssueTestResultAttachmentRuleResponse"
      },
...
      "TestReportedIssueRequest": {
        "properties": {
          "template_id": {
            "type": "string",
            "title": "Template Id",
            "default": ""
          },
          "case_name": {
            "type": "string",
            "title": "Case Name",
            "default": ""
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "url": {
            "type": "string",
            "maxLength": 2083,
            "minLength": 1,
            "format": "uri",
            "title": "Url"
          }
        },
        "type": "object",
        "required": [
          "description",
          "url"
        ],
        "title": "TestReportedIssueRequest"
      },
      "TestReportedIssueResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "template_id": {
            "type": "string",
            "title": "Template Id",
            "default": ""
          },
          "case_name": {
            "type": "string",
            "title": "Case Name",
            "default": ""
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "url": {
            "type": "string",
            "maxLength": 2083,
            "minLength": 1,
            "format": "uri",
            "title": "Url"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "description",
          "url",
          "created_at",
          "updated_at"
        ],
        "title": "TestReportedIssueResponse"
      },
...
      "TestResultStatus": {
        "type": "string",
        "enum": [
          "PASSED",
          "FAILED",
          "SKIPPED"
        ],
        "title": "TestResultStatus"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}