# backend/test_observer/controllers/environments/models.py
# Copyright 2024 Canonical Ltd.
#
# This program is free software: you can redistribute it and/or modify
# it under the terms of the GNU Affero General Public License version 3, as
# published by the Free Software Foundation.
# This program is distributed in the hope that it will be useful,
# but WITHOUT ANY WARRANTY; without even the implied warranty of
# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
# GNU Affero General Public License for more details.
# You should have received a copy of the GNU Affero General Public License
# along with this program.  If not, see <https://www.gnu.org/licenses/>.
#
# SPDX-FileCopyrightText: Copyright 2024 Canonical Ltd.
# SPDX-License-Identifier: AGPL-3.0-only

from datetime import datetime

from pydantic import BaseModel, HttpUrl, model_validator

from test_observer.common.constants import VALID_ISSUE_HOSTS


class EnvironmentReportedIssueRequest(BaseModel):
    environment_name: str
    description: str
    url: HttpUrl | None = None
    is_confirmed: bool

    @model_validator(mode="after")
    def validate_url(self) -> "EnvironmentReportedIssueRequest":
        if self.url is None and self.is_confirmed:
            raise ValueError("A URL is required if the issue is confirmed")

        if self.url is not None and self.url.host not in VALID_ISSUE_HOSTS:
            raise ValueError(f"Issue url must belong to one of {VALID_ISSUE_HOSTS}")

        return self


class EnvironmentReportedIssueResponse(BaseModel):
    id: int
    environment_name: str
    description: str
    url: HttpUrl | None
    is_confirmed: bool
    created_at: datetime
    updated_at: datetime


class EnvironmentsResponse(BaseModel):
    """Response model for environments endpoint"""

    environments: list[str]
    count: int
    limit: int
    offset: int