# backend/test_observer/controllers/issues/shared_models.py
# Copyright 2025 Canonical Ltd.
#
# This program is free software: you can redistribute it and/or modify
# it under the terms of the GNU Affero General Public License version 3, as
# published by the Free Software Foundation.
# This program is distributed in the hope that it will be useful,
# but WITHOUT ANY WARRANTY; without even the implied warranty of
# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
# GNU Affero General Public License for more details.
# You should have received a copy of the GNU Affero General Public License
# along with this program.  If not, see <https://www.gnu.org/licenses/>.
#
# SPDX-FileCopyrightText: Copyright 2025 Canonical Ltd.
# SPDX-License-Identifier: AGPL-3.0-only

from collections.abc import Sequence

from pydantic import (
    AliasPath,
    BaseModel,
    ConfigDict,
    Field,
    HttpUrl,
    field_validator,
)

from test_observer.controllers.execution_metadata.models import ExecutionMetadata
from test_observer.data_access.models import (
    IssueTestResultAttachmentRuleExecutionMetadata,
    TestExecutionMetadata,
)
from test_observer.data_access.models_enums import (
    FamilyName,
    IssueSource,
    IssueStatus,
    TestResultStatus,
)


class MinimalIssueResponse(BaseModel):
    model_config = ConfigDict(from_attributes=True)

    id: int
    source: IssueSource
    project: str
    key: str
    title: str
    status: IssueStatus
    url: HttpUrl
    labels: list[str] | None = None


class MinimalIssueTestResultAttachmentRuleResponse(BaseModel):
    model_config = ConfigDict(from_attributes=True)

    id: int
    enabled: bool

    families: list[FamilyName]
    environment_names: list[str]
    test_case_names: list[str]
    template_ids: list[str]
    test_result_statuses: list[TestResultStatus]
    execution_metadata: ExecutionMetadata

    @field_validator("execution_metadata", mode="before")
    @classmethod
    def convert_execution_metadata(
        cls,
        metadata: Sequence[IssueTestResultAttachmentRuleExecutionMetadata] | ExecutionMetadata,
    ) -> ExecutionMetadata:
        if not isinstance(metadata, ExecutionMetadata):
            return ExecutionMetadata.from_rows(
                [TestExecutionMetadata(category=item.category, value=item.value) for item in metadata]
            )

        return metadata


class MinimalIssueTestResultAttachmentResponse(BaseModel):
    model_config = ConfigDict(from_attributes=True)

    issue: MinimalIssueResponse = Field(validation_alias=AliasPath("issue"))
    attachment_rule: MinimalIssueTestResultAttachmentRuleResponse | None = Field(
        validation_alias=AliasPath("attachment_rule")
    )